refactor(tools): separate string functions from tools.utils
